chore: move load selected fonts to useFont hook, use fonts context
