+++ app/hooks/useFonts.ts
@@ -1,9 +1,11 @@
-import { useEffect, useState } from "react"
+import { useContext, useEffect, useState } from "react"
 
 import { useStorage } from "@plasmohq/storage/hook"
 
+import { loadFont } from "~app/api/fonts"
 import { fetchFontList } from "~app/api/fonts"
 import { STORAGE_KEYS } from "~app/constants"
+import { fontContext } from "~app/context"
 import type { Font, SelectedFonts } from "~app/types"
 
 export const useFonts = () => {
@@ -25,6 +27,24 @@ export const useFonts = () => {
     return () => abortController.abort()
   }, [])
 
+  useEffect(() => {
+    async function loadFonts() {
+      const fonts = Object.values(selectedFont)
+        .flatMap(({ font }) => font)
+        .filter(Boolean)
+
+      try {
+        await Promise.all(fonts.map((font) => loadFont(font)))
+      } catch (err) {
+        console.error("Error loading fonts", err)
+      }
+    }
+
+    if (selectedFont) {
+      loadFonts()
+    }
+  }, [selectedFont])
+
   const setFont = (
     selector: string,
     fontData: { font: Font; fontFaces: string[] }
@@ -36,4 +56,8 @@ export const useFonts = () => {
   }
 
   return { fonts, setFont, selectedFont }
+}
+
+export const useFontsContext = () => {
+  return useContext(fontContext)
 }
